+++ central/jaws_central/rest.py
@@ -153,7 +153,7 @@ def _user_group(user):
     except SQLAlchemyError as error:
         logger.error(error)
         abort(500, {"error": f"Db error; {error}"})
-    current_user.user_group
+    return current_user.user_group
 
 
 def search_runs(user):
@@ -288,7 +288,9 @@ def submit_run(user):
     ):
         # a jaws-site may optionally be restricted to members of a user-group
         user_group = _user_group(user)
-        if user_group != compute_site_config["user_group"] or not _is_admin(user):
+        if user_group == compute_site_config["user_group"] or _is_admin(user):
+            pass
+        else:
             abort(
                 401,
                 {
@@ -349,6 +351,8 @@ def _update_run_status(run, new_status, reason=None):
     """Update run table and insert run_logs entry."""
     status_from = run.status
     run.status = new_status
+    if new_status == "cancelled":
+        run.result = "cancelled"
     try:
         db.session.commit()
     except Exception as error: