+++ test/integration/end-to-end-tests/deployment-tests/test_cancels.py
@@ -15,14 +15,13 @@ import submission_utils as util
 
 WDL = "/WDLs/fq_count.wdl"
 INP = "/test-inputs/fq_count.json"
-FINAL_STATE = "download complete"
-VALID_STATES = ["queued", "running", "download complete"]
+FINAL_STATE = "done"
+VALID_STATES = ["running", "queued", "submitted"]
 CHECK_TRIES = 20000
 CHECK_SLEEP = 5
 
 
-@pytest.mark.parametrize("state", VALID_STATES)
-def test_cancel(dir, site, state):
+def test_cancel(dir, site):
     wdl = dir + WDL
     input_json = dir + INP
     run_id = util.submit_wdl(wdl, input_json, site)["run_id"]
@@ -33,26 +32,24 @@ def test_cancel(dir, site, state):
     # get to the state we are testing
     floating_state = ""
     tries = 0
-    while floating_state != state:
+    while floating_state not in VALID_STATES:
         r_sts, o_sts, e_sts = util.run(status_cmd)
         floating_state = json.loads(o_sts)["status"]
-        if state == FINAL_STATE and floating_state == FINAL_STATE:
-            break
 
         if floating_state == FINAL_STATE:
-            error_message = "This state was not observed: %s" % state
+            error_message = "One of the following states were not seem: %s" % VALID_STATES
             raise Exception(error_message)
 
         if tries < CHECK_TRIES:
             time.sleep(CHECK_SLEEP)
         else:
-            assert 0, "\n**Run took too long; last state: %s. (%s)" % (state, run_id)
+            assert 0, "\n**Run took too long; last state: %s. (%s)" % (floating_state, run_id)
             return
         tries += 1
 
     # submit cancel command to JAWS and gather output
     r_can, o_can, e_can = util.run(cancel_cmd)
-    if state == FINAL_STATE:
+    if floating_state == FINAL_STATE:
         assert r_can == 1, "\n** Return code is not 1. (%s)" % (run_id)
         assert "error" in e_can, "\n** Error keyword absent. (%s)" % (run_id)
         return